% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttester.R
\name{ttester}
\alias{ttester}
\title{Read Mmultiple Student t-Tests Function}
\usage{
ttester(testdata, pairedvalue = FALSE, varvalue = FALSE, pval_sci = FALSE)
}
\arguments{
\item{pairedvalue}{FALSE indicates that the stat::t.test performed will be unpaired, TRUE will be paired (default is FALSE)}

\item{varvalue}{FALSE indicates that equal variance should not be assumed, TRUE assumes equal variance #note in order for the varvalue param to have effect, pairedvalue must be set to FALSE - this choice depends on your study design}

\item{pval_sci}{TRUE if desired output is scientific notation}

\item{data}{The data file that contains the groups that you want to compare as columns or lists}
}
\value{
A list of t.test results and tested_pairs
}
\description{
This function performs multiple t tests between all pairs of columns/lists of a data file and returns a readable output
}
\examples{
ttester(testdata = data.frame(VADeaths), pairedvalue = FALSE, pval_sci = TRUE)
}
